Fix reversion caused by change in graycode variable initialziation
The graycode variable was formerly initialized to 0 only if the mode
was set to 700B.  This was changed with the removal of that mode
to initialize it to 0 for all modes. This caused a reversion that
made at least the 1600 mode produce garbled output. In order to
preserve the previous logic and fix the issue, initialize the
variable to 1.

Fixes: c515387982bb ("removed obsolete 700 and 700B codec 2 modes, fixed unitilaised grey code variable")

Signed-off-by: Annaliese McDermond <nh6z@nh6z.net>